Change JSON format to be more D3 compliant.
I leanred today that D3 has built-in support for a number of hierarchical JSON formats, as long as each level has a 'children' list.

Updated the JSON generation algorithm to support this format.  Parsing the JSON should be a lot easier now.  I almost re-wrote a lot of stuff that D3 had built into it without realizing it.

In short, the format has been changed from this:

    {
      "winner": "?",
      "match_0": {
        "winner": "?",
        "player_0": "One",
        "player_1": "Two"
      },
      "match_1": {
        "winner": "?",
        "player_0": "Thr",
        "player_1": "Fou"
      }
    }

To this:

    {
      "winner": "?",
      "children": [
        {
          "winner": "?",
          "children": [
            {"name": "One"},
            {"name": "Two"}
          ]
        },
        {
          "winner": "?",
          "children": [
            {"name": "Thr"},
            {"name": "Fou"}
          ]
        }
      ]
    }